Add methods to work with ProductUsing

+++ src/main/java/com/taxah/diplomdb/repository/ProductUsingRepository.java
@@ -3,5 +3,5 @@ package com.taxah.diplomdb.repository;
 import com.taxah.diplomdb.model.ProductUsing;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductUsingRepository extends JpaRepository<ProductUsing,Integer> {
+public interface ProductUsingRepository extends JpaRepository<ProductUsing,Long> {
 }
